refactor: primevue languages in a dedicated directory
