+++ src/i18n/primevue/catalan.ts
