+++ src/i18n/primevue/spanish.ts
